Rename store back to jwt-store
In preparation for taking the recoring stuff out of it
